chore: refactoring of component structure
